// front/tfs-front/src/app/entities/permission-type.ts
export enum PermissionType {
  All = "All",
  Read = "Read",
  Write = "Write",
  GrantRead = "GrantRead",
  GrantWrite = "GrantWrite",
  ReadRecursively = "ReadRecursively",
  WriteRecursively = "WriteRecursively",
  GrantReadRecursively = "GrantReadRecursively",
  GrantWriteRecursively = "GrantWriteRecursively"
  //Read = "read",
  //Write = "write",
  //GrantRead = "grant_read",
  //GrantWrite = "grant_write",
  //ReadRecursively = "read_recursively",
  //WriteRecursively = "write_recursively",
  //GrantReadRecursively = "grant_read_recursively",
  //GrantWriteRecursively = "grant_write_recursively"
}